refactor(ui): package structure
